Conflicts:
	samples-test-dev/DemoDB.db
	samples-test-dev/src/main/java/giis/demo/util/SwingMain.java
	samples-test-dev/src/main/java/giis/demo/util/SwingUtil.java

+++ samples-test-dev/src/main/resources/schema.sql
@@ -51,13 +51,13 @@ DROP TABLE IF EXISTS "reservas";
 CREATE TABLE IF NOT EXISTS "reservas" (
 	"idReserva"	INTEGER NOT NULL,
 	"fecha"	TEXT,
-	"horaInicio"	INTEGER,
-	"horaFin"	INTEGER,
+	"horaInicio"	FLOAT,
+	"horaFin"	FLOAT,
 	"idInstalacion"	INTEGER NOT NULL,
 	"idActividad"	INTEGER,
 	"idSocio"	INTEGER,
 	"estado"	INTEGER,
-	PRIMARY KEY("idReserva" AUTOINCREMENT),
+        PRIMARY KEY("idReserva" AUTOINCREMENT),
 	FOREIGN KEY("idSocio") REFERENCES "socios"("idSocio"),
 	FOREIGN KEY("idInstalacion") REFERENCES "instalaciones"("idInstalacion"),
 	FOREIGN KEY("idActividad") REFERENCES "actividades"("idActividad")
